Restructure the code
Make sure that the wizard queries for the all configuration values at
the very beginning so that one can obtain a full configuration file,
when running in dry-run mode.
