+++ pkg/install/v1/install.go
@@ -5,7 +5,6 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"net"
 	"net/url"
 	"strings"
 
@@ -16,14 +15,9 @@ import (
 
 	"github.com/23technologies/23kectl/pkg/common"
 	"github.com/23technologies/23kectl/pkg/logger"
-	"github.com/AlecAivazis/survey/v2"
 	"github.com/fatih/color"
 	"github.com/mitchellh/mapstructure"
 	"github.com/spf13/viper"
-	corev1 "k8s.io/api/core/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
-	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 
@@ -57,8 +51,16 @@ func Install(kubeconfig string, isDryRun bool) error {
 	if err != nil {
 		return err
 	}
-
 	Container.Create = kubeClient.Create
+
+	err = queryConfig(kubeClient)
+	if err != nil {
+		return err
+	}
+	UnmarshalKeConfig(keConfiguration)
+
+	// initialize container
+	// This is espcially important when running in dry run mode
 	if isDryRun {
 		Container.Apply = applyDryRun
 		Container.Create = create
@@ -67,7 +69,7 @@ func Install(kubeconfig string, isDryRun bool) error {
 
 		gitRepoUrl := viper.GetString("admin.gitrepourl")
 		if !strings.Contains(gitRepoUrl, "file://") {
-			return fmt.Errorf("dry run mode only supports local git repositories")
+			return fmt.Errorf("dry run mode only supports local git repositories. I have written a config file for you. If you just wanted to craft an inital config file, you can ignore this error")
 		}
 		gitRepoPath := strings.SplitAfter(gitRepoUrl, "//")[1]
 		_, err = git.PlainInit(gitRepoPath, true)
@@ -75,6 +77,7 @@ func Install(kubeconfig string, isDryRun bool) error {
 			return err
 			}
 	}
+
 
 	fmt.Println("Installing flux")
 	err = installFlux(kubeconfigArgs, kubeclientOptions)
@@ -84,31 +87,6 @@ func Install(kubeconfig string, isDryRun bool) error {
 	}
 
 	err = createBucketSecret(kubeClient)
-	if err != nil {
-		return err
-	}
-
-	err = completeKeConfig(kubeClient)
-	if err != nil {
-		return err
-	}
-
-	err = UnmarshalKeConfig(keConfiguration)
-	if err != nil {
-		return err
-	}
-
-	err = viper.WriteConfig()
-	if err != nil {
-		log.Info("Viper couldn't write config file", "error", err)
-	}
-
-	err = queryAdminConfig()
-	if err != nil {
-		return err
-	}
-
-	err = queryBaseClusterConfig()
 	if err != nil {
 		return err
 	}
@@ -146,18 +124,6 @@ func Install(kubeconfig string, isDryRun bool) error {
 		return err
 	}
 
-	// enable the provider extensions needed for a minimal setup
-	viper.Set("extensionsConfig.provider-"+viper.GetString("baseCluster.provider")+".enabled", true)
-	viper.Set("extensionsConfig."+common.DNS_PROVIDER_TO_PROVIDER[viper.GetString("domainConfig.provider")]+".enabled", true)
-	err = viper.WriteConfig()
-	if err != nil {
-		return err
-	}
-	err = viper.Unmarshal(keConfiguration)
-	if err != nil {
-		return err
-	}
-
 	err = updateConfigRepo(publicKeysConfig)
 	if err != nil {
 		return err
@@ -173,88 +139,6 @@ func Install(kubeconfig string, isDryRun bool) error {
 	return nil
 }
 
-func completeKeConfig(kubeClient client.Client) error {
-	viper.SetDefault("dashboard.sessionSecret", common.RandHex(20))
-	viper.SetDefault("dashboard.clientSecret", common.RandHex(20))
-	viper.SetDefault("kubeApiServer.basicAuthPassword", common.RandHex(20))
-	viper.SetDefault("clusterIdentity", "garden-cluster-"+common.RandHex(5)+"-identity")
-
-	Container.QueryConfigKey("gardenlet.seedPodCidr", func() (any, error) {
-		// If either calico or cilium are used as CNI we can pull the needed info from the cluster
-		// Otherwise prompt the user
-
-		// CALICO
-		// If calico's installed there's an ippool with name "default-ipv4-ippool".
-		ipPool := unstructured.Unstructured{}
-		gvk := schema.GroupVersionKind{
-			Group:   "crd.projectcalico.org",
-			Version: "v1",
-			Kind:    "ippool",
-		}
-		ipPool.SetGroupVersionKind(gvk)
-		err := kubeClient.Get(context.Background(), client.ObjectKey{
-			Namespace: "",
-			Name:      "default-ipv4-ippool",
-		}, &ipPool)
-		if err == nil {
-			return ipPool.Object["spec"].(map[string]interface{})["cidr"].(string), nil
-		}
-
-		// CILIUM
-		// cilium uses a configmap "cilium-config" in the kube-system namespace
-		ciliumConfig := corev1.ConfigMap{}
-		err = kubeClient.Get(context.Background(), client.ObjectKey{
-			Namespace: "kube-system",
-			Name:      "cilium-config",
-		}, &ciliumConfig)
-		if err == nil {
-			return ciliumConfig.Data["cluster-pool-ipv4-cidr"], nil
-		}
-
-		// UNKNOWN
-		// let's prompt for the Pod CIDR
-		prompt := &survey.Input{
-			Message: "Please enter the pod CIDR of your base cluster in the form: x.x.x.x/y",
-		}
-		var queryResult string
-		err = survey.AskOne(prompt, &queryResult, common.WithValidator("required,cidr"))
-		common.ExitOnCtrlC(err)
-		if err != nil {
-			return nil, err
-		}
-		return queryResult, nil
-	})
-
-	if !viper.IsSet("gardenlet.seedServiceCidr") {
-		dummySvc := &corev1.Service{
-			ObjectMeta: metav1.ObjectMeta{
-				Name:      "dummy",
-				Namespace: "default",
-			},
-			Spec: corev1.ServiceSpec{
-				Ports:     []corev1.ServicePort{{Name: "port", Port: 443}},
-				ClusterIP: "1.1.1.1",
-			},
-		}
-		dummyErr := Container.Create(context.Background(), dummySvc)
-		viper.Set("gardenlet.seedServiceCidr", strings.SplitAfter(dummyErr.Error(), "The range of valid IPs is ")[1])
-	}
-
-	if !viper.IsSet("gardener.clusterIP") {
-		seedServiceCidr := viper.GetString("gardenlet.seedServiceCidr")
-		clusterIp, ipnet, _ := net.ParseCIDR(seedServiceCidr)
-
-		// clusterIp[len(clusterIp)-2] += 1
-		clusterIp[len(clusterIp)-1] += 100
-
-		if !ipnet.Contains(clusterIp) {
-			panic(fmt.Sprintf("Your cluster ip (%s) is out of the service IP range: %s", clusterIp, ipnet.String()))
-		}
-		viper.Set("gardener.clusterIP", clusterIp.String())
-	}
-
-	return nil
-}
 
 func getKeConfig() (*KeConfig, error) {
 	keConfig := new(KeConfig)
@@ -268,7 +152,6 @@ func getKeConfig() (*KeConfig, error) {
 
 // unmarshalKeConfig ...
 func UnmarshalKeConfig(config *KeConfig) error {
-
 	err := viper.Unmarshal(config)
 	if err != nil {
 		return err
@@ -295,8 +178,9 @@ func UnmarshalKeConfig(config *KeConfig) error {
 }
 
 // create ...
+// implements a dry run create function. It only outputs the objects
+// which would be created in the cluster
 func create(ctx context.Context, obj client.Object, opts ...client.CreateOption) error {
-
 	tmp, err := json.Marshal(obj)
 	jsonReader := bytes.NewReader(tmp)
 	yamlWriter := strings.Builder{}
