+++ pkg/install/v1/secrets.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 
 	"github.com/23technologies/23kectl/pkg/common"
-	"github.com/AlecAivazis/survey/v2"
 	"github.com/spf13/viper"
 	corev1 "k8s.io/api/core/v1"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -45,29 +44,6 @@ func createBucketSecret(kubeClient client.Client) error {
 }
 
 func create23keConfigSecret(kubeClient client.Client) error {
-	Container.QueryConfigKey("issuer.acme.email", func() (any, error) {
-		prompt := &survey.Input{
-			Message: "Please enter your email address for acme certificate generation",
-			Default: viper.GetString("admin.email"),
-		}
-		var queryResult string
-		err := survey.AskOne(prompt, &queryResult, common.WithValidator("required,email"))
-		common.ExitOnCtrlC(err)
-		if err != nil {
-			return nil, err
-		}
-
-		return queryResult, nil
-	})
-
-	Container.QueryConfigKey("domainConfig", func() (any, error) {
-		domainConfig, err := queryDomainConfig()
-		if err != nil {
-			return nil, err
-		}
-		return domainConfig, nil
-	})
-
 	fmt.Println("Creating '23ke-config' secret")
 
 	tpl, err := makeTemplate().Parse(`
